+++ plugin-packs/postcss-preset-env/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Preset Env
 
+### Unreleased (minor)
+
+- Added `@csstools/postcss-gradients-interpolation-method` [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-gradients-interpolation-method#readme) for usage details.
+- Updated `cssdb` to `7.5.2` (patch)
+
 ### 8.1.0 (March 25, 2023)
 
 - Added `@csstools/postcss-color-mix-function` [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-color-mix-function#readme) for usage details.